Explaination of Sales, Profit and Loss, Balance Sheet, and Accounting Ratios.
This report consists of a detailed financial analysis of the company for 2018, 2019, and 2020. It 
also provides some base graphical analytics of the financial performance and financial position 
on a country basis (Canada, China, Denmark, France, India, Netherlands, and New Zealand)
This dashboard incorporated dynamic elements like slicers, cards, graphs, and maps to depict the 
company's performance over the abovementioned period.